openfst-1.4.0.tar.gz r1
SHA-256: eb557f37560438f03912b4e43335c4c9e72aa486d4f2046127131185eb88f17a

+++ src/lib/flags.cc
@@ -49,7 +49,7 @@ void SetFlags(const char *usage, int *argc, char ***argv,
     string val = "";
 
     // split argval (arg=val) into arg and val
-    int pos = argval.find("=");
+    size_t pos = argval.find("=");
     if (pos != string::npos) {
       arg = argval.substr(0, pos);
       val = argval.substr(pos + 1);
